requirements.txt
# AI Email Campaign Manager - Dependencies
# The most INSANE AI-powered email marketing system ever built! 🔥

# Core Framework
Flask==3.1.2
Werkzeug==3.1.3

# Data Processing
pandas==2.3.2
numpy==2.0.2
openpyxl==3.1.5

# Environment Management
python-dotenv==1.1.1

# HTTP Requests
requests==2.31.0

# Email Processing
# smtplib and imaplib are built-in Python modules
# No additional packages needed for email functionality

# AI Integration
# Gemini AI integration via requests (no additional packages needed)

# Development Dependencies (Optional)
# pytest==7.4.0
# black==23.7.0
# flake8==6.0.0

# Production Dependencies (Optional)
# gunicorn==21.2.0
# waitress==2.1.2

# Security Dependencies (Optional)
# cryptography==41.0.4
# bcrypt==4.0.1

# Monitoring Dependencies (Optional)
# psutil==5.9.5
# schedule==1.2.0

# Note: This system is designed to be lightweight and efficient
# All core functionality works with the packages listed above
# Additional packages can be added based on specific requirements